+++ MK4duo/src/lcd/nextionlcd/nextionlcd.cpp
@@ -97,15 +97,14 @@
 
   #endif
 
-  bool        NextionON                   = false,
-              show_Wave                   = true;
+  bool        NextionON                   = false;
   uint8_t     PageID                      = 0;
   uint16_t    slidermaxval                = 20;
   char        buffer[NEXTION_BUFFER_SIZE] = { 0 };
 
   #if HAS_SD_SUPPORT
-    // 0 card not present, 1 SD not insert, 2 SD insert, 3 SD printing
-    enum SDstatus_enum {NO_SD = 0, SD_NO_INSERT = 1, SD_INSERT = 2, SD_PRINTING = 3, SD_PAUSE = 4 };
+    // 0 card not present, 1 SD not insert, 2 SD insert, 3 SD printing, 4 SD paused, 5 USB printing
+    enum SDstatus_enum {NO_SD = 0, SD_NO_INSERT = 1, SD_INSERT = 2, SD_PRINTING = 3, SD_PAUSE = 4, USB_PRINTING = 5 };
     SDstatus_enum SDstatus    = NO_SD;
     NexUpload Firmware(NEXTION_FIRMWARE_FILE, 57600);
   #endif
@@ -572,17 +571,8 @@
   }
 
   static void degtoLCD(const uint8_t h, float temp) {
-
     NOMORE(temp, 999);
-
     nexlcd.setValue(*heater_list0[h], temp);
-
-    #if ENABLED(NEXTION_GFX)
-      if (!printer.isPrinting() && !nexlcd.getObjVis(Wavetemp) && show_Wave) {
-        nexlcd.SetVisibility(Wavetemp, true);
-      }
-    #endif
-
   }
 
   static void targetdegtoLCD(const uint8_t h, const float temp) {
@@ -738,28 +728,38 @@
           PreviouspercentDone = printer.progress;
         }
 
-        #if HAS_SD_SUPPORT
-
-          if (card.isFileOpen()) {
-            if (IS_SD_PRINTING() && SDstatus != SD_PRINTING) {
+        if (IS_SD_FILE_OPEN()) {
+          if (IS_SD_PRINTING()) {
+            if (SDstatus != SD_PRINTING) {
               SDstatus = SD_PRINTING;
               nexlcd.setValue(SD, SDstatus);
             }
-            else if (!IS_SD_PRINTING() && SDstatus != SD_PAUSE) {
+          }
+          else {
+            if (SDstatus != SD_PAUSE) {
               SDstatus = SD_PAUSE;
               nexlcd.setValue(SD, SDstatus);
             }
           }
-          else if (card.isOK() && SDstatus != SD_INSERT) {
+        }
+        else if (print_job_counter.isRunning()) {
+          if (SDstatus != USB_PRINTING) {
+            SDstatus = USB_PRINTING;
+            nexlcd.setValue(SD, SDstatus);
+          }
+        }
+        else if (IS_SD_OK()) {
+          if (SDstatus != SD_INSERT) {
             SDstatus = SD_INSERT;
             nexlcd.setValue(SD, SDstatus);
           }
-          else if (!card.isOK() && SDstatus != SD_NO_INSERT) {
+        }
+        else if (!IS_SD_OK()) {
+          if (SDstatus != SD_NO_INSERT) {
             SDstatus = SD_NO_INSERT;
             nexlcd.setValue(SD, SDstatus);
           }
-
-        #endif // HAS_SD_SUPPORT
+        }
 
         break;
 
@@ -788,12 +788,9 @@
       gfx.set_scale(scale);
     }
 
-    void gfx_clear(const float x, const float y, const float z, bool force_clear) {
-      if (PageID == 2 && (printer.isPrinting() || force_clear)) {
-        nexlcd.SetVisibility(Wavetemp, false);
-        show_Wave = !force_clear;
+    void gfx_clear(const float x, const float y, const float z) {
+      if (PageID == 2 && printer.isPrinting())
         gfx.clear(x, y, z);
-      }
     }
 
     void gfx_cursor_to(const float x, const float y, const float z, bool force_cursor) {